+++ pytsite/auth/_eh.py
@@ -34,7 +34,7 @@ def app_setup():
 
         admin_user = _functions.create_user(email)
         admin_user.f_set('first_name', _lang.t('pytsite.auth@administrator'))
-        admin_user.f_set('nickname', _util.transform_str_1(admin_user.full_name))
+        admin_user.f_set('nickname', _util.transform_str_2(admin_user.full_name))
         admin_user.f_add('roles', _functions.get_role('admin'))
         admin_user.save()
         _console.print_success(_lang.t('pytsite.auth@user_has_been_created', {'login': admin_user.f_get('login')}))
@@ -60,5 +60,5 @@ def _update_0_13():
     from pytsite import util
     for user in _functions.find_users(False).get():
         if not user.nickname:
-            user.f_set('nickname', util.transform_str_1(user.full_name)).save()
+            user.f_set('nickname', util.transform_str_2(user.full_name)).save()
             _console.print_info('User updated: {}'.format(user.login))