+++ pytsite/route_alias/_model.py
@@ -1,10 +1,10 @@
 """Route Alias Model.
 """
+from pytsite import odm as _odm
+
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
-
-from pytsite import odm as _odm
 
 
 class RouteAlias(_odm.Model):